Title: 실수 입력받아 계산, Time: 140ms, Memory: 11MB, Status: Passed - CodeTree

+++ 241106/실수 입력받아 계산/calculation-by-inputting-a-real-number.java
@@ -4,8 +4,8 @@ public class Main {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
 
-        double a = sc.nextInt();
-        double b = sc.nextInt();
+        double a = sc.nextDouble();
+        double b = sc.nextDouble();
 
         double c = a + b;
